TestsInfrastructure project created, reusable items moved there
