+++ src/Tests/CaptainHook.Tests/CaptainHook.Tests.csproj
@@ -44,6 +44,7 @@
     <ProjectReference Include="..\..\CaptainHook.Domain\CaptainHook.Domain.csproj" />
     <ProjectReference Include="..\..\CaptainHook.EventHandlerActor\CaptainHook.EventHandlerActor.csproj" />
     <ProjectReference Include="..\..\CaptainHook.EventReaderService\CaptainHook.EventReaderService.csproj" />
+    <ProjectReference Include="..\CaptainHook.TestsInfrastructure\CaptainHook.TestsInfrastructure.csproj" />
   </ItemGroup>
 
 </Project>
